requirements.txt
asgiref==3.8.1
Django==4.2.15
django-phonenumber-field==8.0.0
phonenumbers==8.13.43
sqlparse==0.5.1
typing_extensions==4.12.2
